qb17.java
The abstract Vegetable class has three subclasses named Potato, Brinjal and Tomato. Write 
an application that demonstrates how to establish this class hierarchy. Declare one instance 
variable of type String that indicates the color of a vegetable. Create and display instances of 
these objects. Override the toString() method of Object to return a string with the name of 
the vegetable and its color.